pyproject.toml
[project]
name = "napy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Morten Hansen", email = "morten@winterop.com" }]
requires-python = ">=3.13"
dependencies = [
    "ansi2html>=1.9.2",
    "pydantic-settings>=2.12.0",
    "python-daemon>=3.0.1",
    "python-telegram-bot>=21.0",
]

[project.scripts]
napy = "napy:main_entry_point"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
